+++ pages/api/municipalities.tsx
@@ -1,4 +1,5 @@
 import type { NextApiRequest, NextApiResponse } from 'next'
+import { getMunicipalities } from '../../utils/emissionService';
 
 export default function userHandler(req: NextApiRequest, res: NextApiResponse) {
     const {
@@ -8,8 +9,18 @@ export default function userHandler(req: NextApiRequest, res: NextApiResponse) {
   
     switch (method) {
       case 'GET':
-        // Get data from your database
-        res.status(200).json({ id, name: `Alla kommuner` })
+        
+          getMunicipalities().then((municipalities) => {
+            if(municipalities.length < 1) {
+              res.status(404).json("Inga kommuner hittades")
+            }
+  
+            res.status(200).json(municipalities)
+          }).catch((error) => {
+            res.status(500).json(error)
+          })
+          
+
         break
       // case 'PUT':
       //   // Update or create data in your database
